Move android_config.dart to constants folder
