+++ lib/src/util/file/android_manager.dart
@@ -1,7 +1,7 @@
 import 'dart:io' show Directory;
 
 import '../../models/exceptions/android_build_exception.dart';
-import '../../config/android_config.dart';
+import '../../constants/android_config.dart';
 import '../../models/file_manager.dart';
 
 class AndroidManager {
